requirements.txt
# General 
pyyaml==6.0.0
pydantic==2.7.3
openpyxl==3.1.5

# HuggingFace
datasets==2.14.0

# Feature Extraction
sentence-transformers==3.4.0
textstat==0.7.5

# Inference
transformers==4.50.2
torch==2.6.0
accelerate==1.5.2
sentencepiece==0.2.0
protobuf==6.30.2

# Data Connectors
psycopg2-binary==2.9.6      
redis==4.6.0

# Data Processing
numpy==1.24.3
pandas==2.0.0

# Measurements
pynvml==12.0.0

# API
uvicorn==0.34.0
fastapi==0.115.11

# Visualization
plotly==5.14.0
seaborn==0.12.2
matplotlib==3.7.1

# Analysis
scipy==1.11.1
scikit-learn==1.5.2
statsmodels==0.14.1

# Logging
loguru==0.7.3
